+++ src/main/java/neoflex/task/vacationcalculator/service/interfaces/CalculateVacationPayoutService.java
@@ -5,6 +5,6 @@ import neoflex.task.vacationcalculator.dto.CalculateVacationResponseDTO;
 import java.time.LocalDate;
 
 public interface CalculateVacationPayoutService {
-    CalculateVacationResponseDTO calculateVacationPayout(float avgSalary, int numberOfDays);
-    CalculateVacationResponseDTO calculateVacationPayout(float avgSalary, LocalDate startDate, LocalDate endDate);
+    CalculateVacationResponseDTO calculateVacationPayout(double avgSalary, int numberOfDays);
+    CalculateVacationResponseDTO calculateVacationPayout(double avgSalary, LocalDate startDate, LocalDate endDate);
 }
